Move rewrite_union to goto-programs
There is nothing symex-specific about it, and it is safely usable by code that
does not depend on symex. Resolves some known-inappropriate module dependencies.
