+++ src/goto-diff/module_dependencies.txt
@@ -4,7 +4,6 @@ cpp
 goto-diff
 goto-instrument
 goto-programs
-goto-symex # dubious - rewrite_union and adjust_float should be moved to goto-programs
 java_bytecode # will go away
 langapi # should go away
 pointer-analysis
