+++ src/goto-programs/Makefile
@@ -50,6 +50,7 @@ SRC = adjust_float_expressions.cpp \
       remove_unused_functions.cpp \
       remove_vector.cpp \
       remove_virtual_functions.cpp \
+      rewrite_union.cpp \
       replace_calls.cpp \
       resolve_inherited_component.cpp \
       safety_checker.cpp \
